Move `UIViewController` snapshot helpers to separate file
